Update readme with fixes to grammar, as well as add instructions to add an "update(entry: ...)" function that was missing. Without it, the user would not be able to have the changes to their entries persist.
Update addEntry function to take in the parameters needed to initialize an Entry, instead of having the Entry be initialized in a view controller.

Add an "update(entry: ...)" function in the EntryController to match the instructions.

Update saveButtonTapped IBAction to safely unwrap the text from the text field and view, and now calls the update(entry: ..." function.

The "updateViews" function on the EntryDetailViewController was crashing the app when being called in the didSet of the entry property. It appears the view isn't loaded prior to calling the didSet. When the function was called, it would crash because the textFields were nil. Now, it will only be called in the didSet if the view is already loaded. Otherwise, the viewDidLoad will call it.